requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
msgpack==1.0.3
neovim==0.3.1
outcome==1.1.0
pycparser==2.21
pynvim==0.4.3
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.27.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
spotipy
python-dotenv
chromedriver-autoinstaller
